[Networking] Search for CA roots if libcurl doesn't know where they are.
Normally, distro maintainers will tell libcurl where to look when they
build it, and up to now we've been relying on that.  That doesn't work
for the fully static Linux build, where we're building our own libcurl,
and where the idea is that we'll run on any old Linux system.

To make TLS work under that circumstance, we'll need to look in a few
likely places for CA root files.  We only do this if libcurl doesn't
already know where to look.

rdar://123434144